+++ text/freetype_font.h
@@ -1,5 +1,5 @@
 // LAF Text Library
-// Copyright (C) 2020-2024  Igara Studio S.A.
+// Copyright (C) 2020-2025  Igara Studio S.A.
 // Copyright (C) 2016-2017  David Capello
 //
 // This file is released under the terms of the MIT license.
@@ -33,7 +33,7 @@ public:
                     gfx::RectF* bounds,
                     const os::Paint* paint) const override;
   bool isScalable() const override;
-  void setSize(int size) override;
+  void setSize(float size) override;
   bool antialias() const override;
   void setAntialias(bool antialias) override;
