+++ text/skia_font.cpp
@@ -1,5 +1,5 @@
 // LAF Text Library
-// Copyright (c) 2024  Igara Studio S.A.
+// Copyright (c) 2024-2025  Igara Studio S.A.
 //
 // This file is released under the terms of the MIT license.
 // Read LICENSE.txt for more information.
@@ -92,7 +92,7 @@ bool SkiaFont::isScalable() const
   return true;
 }
 
-void SkiaFont::setSize(int size)
+void SkiaFont::setSize(float size)
 {
   m_skFont.setSize(size);
 }
