+++ text/skia_font.h
@@ -1,5 +1,5 @@
 // LAF Text Library
-// Copyright (c) 2024  Igara Studio S.A.
+// Copyright (c) 2024-2025  Igara Studio S.A.
 //
 // This file is released under the terms of the MIT license.
 // Read LICENSE.txt for more information.
@@ -29,7 +29,7 @@ public:
                     gfx::RectF* bounds,
                     const os::Paint* paint) const override;
   bool isScalable() const override;
-  void setSize(int size) override;
+  void setSize(float size) override;
   bool antialias() const override;
   void setAntialias(bool antialias) override;
